+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Extensions/SDK6DataMappingExtensions.asmdef
@@ -3,7 +3,10 @@
     "rootNamespace": "",
     "references": [
         "GUID:3c7b57a14671040bd8c549056adc04f5",
-        "GUID:0b167d3fa6c88454ea3eb69cbabe4eb7"
+        "GUID:0b167d3fa6c88454ea3eb69cbabe4eb7",
+        "GUID:132013f160565434195ba5960832b346",
+        "GUID:a881b57670b1d2747a6d7f9e32b63230",
+        "GUID:3047fd2f05bcaae498a99a3f9321f9a8"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
